+++ src/SbeCodeGenerator/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SbeCodeGenerator.Tests")]
+[assembly: InternalsVisibleTo("SbeCodeGenerator.Profiling")]